The problem is with panel.
When updating the figure, panel is writing the image as .png and then reading the info. In this step of reading and/or creating is where it fails to do so and cannot update the figure.
go to: projector.trigger()

problem is in the pn.pane.Matplotlib class, the method _img(), function self.object.canvas.print_figure(b, bbox_inches=bbox_inches)